+++ CHANGES.rst
@@ -18,6 +18,9 @@ dev (master)
 * ``HTTPResponse`` contains the last ``Retry`` object, which now also
   contains retries history. (Issue #848)
 
+* Timeout can no longer be set as boolean, and must be greater than zero. 
+  (PR #924)
+
 * ... [Short description of non-trivial change.] (Issue #)